+++ examples/advection_diffusion/vanilla_sdc.cpp
@@ -16,11 +16,11 @@ int main(int argc, char **argv)
 {
   pfasst::SDC sdc;
 
-  const int    nsteps = 4;
+  const size_t nsteps = 4;
   const double dt     = 0.01;
-  const int    nnodes = 5;
-  const int    ndofs  = 64;
-  const int    niters = 4;
+  const size_t nnodes = 5;
+  const size_t ndofs  = 64;
+  const size_t niters = 4;
 
   auto  nodes   = pfasst::compute_nodes(nnodes, "gauss-lobatto");
   auto* factory = new pfasst::encap::VectorFactory<double>(ndofs);
